# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Supermarket API Documentation",
        "description": "Documentation for Supermarket Departments API",
        "contact": {
            "name": "API Support",
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products with optional filtering",
                "operationId": "34ee6e9e8fa0c95d9bd89cfc374a81e2",
                "parameters": [
                    {
                        "name": "department_id",
                        "in": "query",
                        "description": "Filter by department ID",
                        "required": false
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "promotional",
                        "in": "query",
                        "description": "Filter promotional products",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by product name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Smartphone"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Latest model smartphone"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 699.99
                                                    },
                                                    "stock_quantity": {
                                                        "type": "integer",
                                                        "example": 50
                                                    },
                                                    "min_stock_threshold": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "smartphone"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Phones"
                                                    },
                                                    "is_promotional": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "department_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "operationId": "ef2288e0d31fb52d40e1ad1fd5e1ef8b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "stock_quantity",
                                    "min_stock_threshold",
                                    "department_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Smartphone"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Latest model smartphone"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 699.99
                                    },
                                    "stock_quantity": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "min_stock_threshold": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Phones"
                                    },
                                    "is_promotional": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Smartphone"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Latest model smartphone"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 699.99
                                                },
                                                "stock_quantity": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "min_stock_threshold": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "smartphone"
                                                },
                                                "category": {
                                                    "type": "string",
                                                    "example": "Phones"
                                                },
                                                "is_promotional": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "department_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product details",
                "operationId": "0d2750d3432c0a2166a357832496e239",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "operationId": "45449c51f7a358aabbb1c8e4966c293c",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "operationId": "1813b2a21fdf00f6f7bf871475e284e4",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/departments/{department}/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products for a specific department",
                "operationId": "d3b7e3bcbd526dd52e15e451f98e0afa",
                "parameters": [
                    {
                        "name": "department",
                        "in": "path",
                        "description": "Department ID",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Department not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}